create thunks and move to their own file
